organize capabilities loader for current test only
