+++ carina-webdriver/src/main/java/com/qaprosoft/carina/core/foundation/webdriver/core/capability/CapabilitiesLoader.java
@@ -43,6 +43,19 @@ public class CapabilitiesLoader {
      *            String path to the properties file with custom capabilities and properties
      */
     public void loadCapabilities(String fileName) {
+        loadCapabilities(fileName, false);
+    }
+    
+    /**
+     * Load capabilities and properties from external file into the global or current test CONFIG context.
+     * {@code capabilities.<name>=<value> will be attached to each WebDriver capabilities
+     * <name>=<value> will override appropriate configuration parameter by new <value>}
+     *  
+     * @param fileName
+     *            String path to the properties file with custom capabilities and properties
+     * @param currentTestOnly boolean
+     */
+    public void loadCapabilities(String fileName, boolean currentTestOnly) {
         LOGGER.info("Loading capabilities to global context from " + fileName);
         Properties props = loadProperties(fileName);
 
@@ -53,7 +66,7 @@ public class CapabilitiesLoader {
             String key = entry.getKey();
             LOGGER.info("Set custom property: " + key + "; value: " + value);
             // add each property directly into CONFIG
-            R.CONFIG.put(key, value);
+            R.CONFIG.put(key, value, currentTestOnly);
         }
 
     }
